add exists to check globbing

+++ src/ci/agent.sh
@@ -5,6 +5,10 @@
 
 set -ex
 
+exists() {
+    [ -e "$1" ]
+}
+
 export RUSTC_WRAPPER=$(which sccache)
 
 mkdir -p artifacts/agent
@@ -27,6 +31,8 @@ cargo test --release --manifest-path ./onefuzz/Cargo.toml
 cp target/release/onefuzz-agent* ../../artifacts/agent
 cp target/release/onefuzz-supervisor* ../../artifacts/agent
 
-for file in target/release/*.pdb; do
-    cp ${file} ../../artifacts/agent
-done
+if exists target/release/*.pdb; then
+    for file in target/release/*.pdb; do
+        cp ${file} ../../artifacts/agent
+    done
+fi
